# api/openapi.yaml
openapi: 3.1.0
info:
  title: WorkBuddy Expert Team OS
  version: 1.0.0
paths:
...
  /v1/agent-runs/{run_id}/execute:
    post:
      summary: Execute Run
      operationId: execute_run_v1_agent_runs__run_id__execute_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecuteIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/skills/{release_id}/test:
    post:
      summary: Test Skill
      operationId: test_skill_v1_skills__release_id__test_post
      parameters:
      - name: release_id
        in: path
        required: true
        schema:
          type: string
          title: Release Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillTestIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/model/invocations:
    get:
      summary: Model Invocations
      operationId: model_invocations_v1_model_invocations_get
      parameters:
      - name: task_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Task Type
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/dispatch/{decision_id}/feedback:
    post:
      summary: Add Dispatch Feedback
      operationId: add_dispatch_feedback_v1_dispatch__decision_id__feedback_post
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchFeedbackIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pilot/sync-runs:
    get:
      summary: Sync Runs
      operationId: sync_runs_v1_pilot_sync_runs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/policies/external-email:
    get:
      summary: Get External Email Policy
      operationId: get_external_email_policy_v1_policies_external_email_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    put:
      summary: Update External Email Policy
      operationId: update_external_email_policy_v1_policies_external_email_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantPolicyIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/operations/{operation_id}/attempts:
    get:
      summary: Operation Attempts
      operationId: operation_attempts_v1_operations__operation_id__attempts_get
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: string
          title: Operation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/connectors/graph/start:
    get:
      summary: Graph Start
      operationId: graph_start_v1_connectors_graph_start_get
      parameters:
      - name: enable_send
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Enable Send
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/connectors/graph/callback:
    get:
      summary: Graph Callback
      operationId: graph_callback_v1_connectors_graph_callback_get
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
          title: Code
      - name: state
        in: query
        required: true
        schema:
          type: string
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/connectors/graph/accounts/{account_id}/sync:
    post:
      summary: Graph Sync
      operationId: graph_sync_v1_connectors_graph_accounts__account_id__sync_post
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      - name: folder_id
        in: query
        required: false
        schema:
          type: string
          default: inbox
          title: Folder Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/connectors/graph/accounts/{account_id}/watch:
    post:
      summary: Graph Watch
      operationId: graph_watch_v1_connectors_graph_accounts__account_id__watch_post
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/connectors/graph/webhook:
    get:
      summary: Graph Webhook Validation
      operationId: graph_webhook_validation
      parameters:
      - name: validationToken
        in: query
        required: true
        schema:
          type: string
          title: Validationtoken
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Graph Webhook Receive
      operationId: graph_webhook_receive
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...
components:
  schemas:
    AgentExecuteIn:
      properties:
        force_provider:
          type: string
          const: configured
          title: Force Provider
          default: configured
      additionalProperties: false
      type: object
      title: AgentExecuteIn
...
    DispatchFeedbackIn:
      properties:
        confirmed_team_key:
          type: string
          title: Confirmed Team Key
        corrected_risk_level:
          anyOf:
          - type: string
            enum:
            - low
            - medium
            - high
            - critical
          - type: 'null'
          title: Corrected Risk Level
        comment:
          type: string
          maxLength: 2000
          title: Comment
          default: ''
      additionalProperties: false
      type: object
      required:
      - confirmed_team_key
      title: DispatchFeedbackIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SkillTestIn:
      properties:
        test_input:
          additionalProperties: true
          type: object
          title: Test Input
          default: {}
      additionalProperties: false
      type: object
      title: SkillTestIn
...
    TenantPolicyIn:
      properties:
        config:
          additionalProperties: true
          type: object
          title: Config
        expected_version:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Expected Version
      additionalProperties: false
      type: object
      required:
      - config
      title: TenantPolicyIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...